+++ src/engine/context.ts
@@ -1,0 +1,17 @@
+import { GameState } from "./state";
+import { Services, PassiveManager } from "./services";
+import { Registry } from "./registry";
+import type { ActionDef } from "./actions";
+import type { BuildingDef } from "./buildings";
+
+export class EngineContext {
+  constructor(
+    public game: GameState,
+    public services: Services,
+    public actions: Registry<ActionDef>,
+    public buildings: Registry<BuildingDef>,
+    public passives: PassiveManager,
+  ) {}
+  get me() { return this.game.active; }
+  get opp() { return this.game.opponent; }
+}
